+++ zchecker/cli_helpers.py
@@ -2,7 +2,10 @@
 """Command-line helpers."""
 import re
 import os
+import astropy.units as u
 from astropy.time import Time
+from sbpy.data import Orbit
+from sbsearch.util import epochs_to_jd, epochs_to_time
 
 __all__ = ['object_list', 'as_time']
 
@@ -24,3 +27,30 @@ def as_time(date):
             raise ValueError(
                 'Bad date: {}; date format is YYYY-MM-DD.'.format(date))
     return Time(date)
+
+def args2orbit(args):
+    try:
+        Tp = float(args.Tp)
+    except ValueError:
+        Tp = args.Tp
+
+    try:
+        epoch = float(args.epoch)
+    except ValueError:
+        epoch = args.epoch
+
+    orbit = Orbit.from_dict({
+        'M': args.M,
+        'K': args.K,
+        'H': 5,
+        'G': 0.15,
+        'q': args.q * u.au,
+        'e': args.e,
+        'incl': args.i * u.deg,
+        'Omega': args.node * u.deg,
+        'w': args.argperi * u.deg,
+        'Tp_jd': epochs_to_time([Tp], scale='tt').jd * u.day
+    })
+    orbit['orbittype'] = ['COM']
+    orbit['epoch'] = epochs_to_time([epoch], scale='tt')
+    return orbit