+++ zchecker/zchecker.py
@@ -165,6 +165,7 @@ class ZChecker(SBSearch):
         n_rows = self.executemany('DELETE FROM ztf_cutouts WHERE foundid=?',
                                   [[i] for i in foundids])
         n_files = self.clean_stale_files()
+        self.db.commit()
         return n_rows, n_files
 
     def clean_stacks(self, stackids):
@@ -185,6 +186,7 @@ class ZChecker(SBSearch):
         n_rows = self.executemany('DELETE FROM ztf_stacks WHERE stackid=?',
                                   [[i] for i in stackids])
         n_files = self.clean_stale_files()
+        self.db.commit()
         return n_rows, n_files
 
     def clean_stale_files(self):
@@ -211,6 +213,7 @@ class ZChecker(SBSearch):
             self.db.executemany(
                 'DELETE FROM ztf_stale_files WHERE rowid=?', rowids)
             self.logger.info('{} stale archive files removed.'.format(count))
+        self.db.commit()
         return count
 
     def download_cutouts(self, objects=None, clean_failed=True,
@@ -238,10 +241,7 @@ class ZChecker(SBSearch):
         fntemplate = ('{desgfile}/{desgfile}-{datetime}-{prepost}{rh:.3f}'
                       '-{filtercode[1]}-ztf.fits')
 
-        cmd = '''SELECT * FROM ztf_found
-        INNER JOIN obj USING (objid)
-        LEFT JOIN ztf_cutouts USING (foundid)
-        '''
+        cmd = 'SELECT foundid FROM found LEFT JOIN ztf_cutouts USING (foundid)'
 
         constraints = []
 
@@ -266,8 +266,7 @@ class ZChecker(SBSearch):
 
         cmd, parameters = util.assemble_sql(cmd, [], constraints)
         count = self.db.execute(
-            cmd.replace(' * ', ' COUNT() ', 1), parameters).fetchone()[0]
-        rows = self.db.execute(cmd, parameters)
+            cmd.replace(' foundid ', ' COUNT() ', 1), parameters).fetchone()[0]
 
         if count == 0:
             self.logger.info('No cutouts to download.')
@@ -280,11 +279,17 @@ class ZChecker(SBSearch):
 
         self.logger.info('{} {} cutouts.'.format(verb, count))
 
+        foundids = self.db.iterate_over(cmd, parameters)
         missing = 0
         downloaded = 0
         with ztf.IRSA(path, self.config.auth) as irsa:
-            for row in util.iterate_over(rows):
-                row = OrderedDict(row)
+            for foundid in foundids:
+                row = OrderedDict(self.db.execute('''
+                SELECT * FROM found
+                INNER JOIN ztf USING (obsid)
+                INNER JOIN obj USING (objid)
+                WHERE foundid=:foundid
+                ''', {'foundid': foundid[0]}).fetchone())
                 count -= 1
 
                 if missing_files:
@@ -321,6 +326,111 @@ class ZChecker(SBSearch):
 
         self.logger.info('{} files successfully downloaded.'
                          .format(downloaded))
+
+    def find_by_orbit(self, desg, orbit, start=None, stop=None, download=None):
+        """Search for object by orbital elements.
+
+        Parameters
+        ----------
+        orbit : `~sbpy.data.orbit.Orbit`
+            Target name (`'desg'`) and orbital parameters.  If `'M'`
+            and `'K'` are provided, they will be used for the apparent
+            magnitude model.
+
+        start : float or `~astropy.time.Time`, optional
+            Search after this epoch, inclusive.
+
+        stop : float or `~astropy.time.Time`, optional
+            Search before this epoch, inclusive.
+
+        download : string, optional
+            Download 'cutouts' or 'fullframe'.
+
+        Returns
+        -------
+        tab : `~astropy.table.Table` or ``None``
+
+        """
+
+        tab = super().find_by_orbit(orbit, start=start, stop=stop)
+
+        if download is None:
+            return tab
+
+        if tab is None:
+            self.logger.info('Nothing to download.')
+            return None
+
+        rows = tab.copy()
+
+        # add ZData required meta data
+        undef = ['undef'] * len(rows)
+        na = ['N/A'] * len(rows)
+        rows['desg'] = [desg] * len(rows)
+        rows['objid'] = na
+        rows['obsjd'] = Time(rows['date']).jd
+        rows['phase'] = undef
+        rows['rdot'] = np.zeros(len(rows))
+        rows['sangle'] = undef
+        rows['vangle'] = undef
+        rows['trueanomaly'] = undef
+        rows['tmtp'] = undef
+        rows['ra'] = rows['RA']
+        rows['dec'] = rows['Dec']
+        rows['dra'] = undef
+        rows['ddec'] = undef
+        rows['ra3sig'] = undef
+        rows['dec3sig'] = undef
+        rows['foundid'] = na
+        rows['ccdid'] = rows['ccd']
+        rows['qid'] = rows['quad']
+        rows['filtercode'] = rows['filter']
+
+        path = self.config['cutout path']
+        fntemplate = ('found-by-orbit/{desgfile}/{desgfile}-{datetime}-{rh:.3f}'
+                      '-{filtercode[1]}-ztf.fits')
+        if download is 'cutouts':
+            size = self.config['cutout size']
+        elif download is 'fullframe':
+            size = None
+        else:
+            raise ValueError('download must be cutout or fullframe.')
+
+        count = len(rows)
+        exists = 0
+        self.logger.info('Checking for {} images.'.format(count))
+
+        with ztf.IRSA(path, self.config.auth) as irsa:
+            for i in range(len(rows)):
+                row = {}
+                for col in rows.colnames:
+                    row[col] = rows[col][i]
+
+                try:
+                    with ZData(irsa, path, fntemplate, self.logger,
+                               preserve_case=True, **row) as cutout:
+                        count -= 1
+
+                        if os.path.exists(cutout.fn):
+                            exists += 1
+                            continue
+
+                        cutout.append('sci', size=size)
+
+                        for img in ['mask', 'psf', 'diff', 'ref']:
+                            try:
+                                cutout.append(img, size=size)
+                            except ZCheckerError:
+                                pass
+                except ZCheckerError as e:
+                    self.logger.error('{} - {}'.format(cutout.fn, str(e)))
+
+                self.logger.debug('  [{}] {}'.format(count + 1, cutout.fn))
+
+        self.logger.info('{} downloaded, {} already exist{}.'.format(
+            len(rows) - exists, exists, 's' if exists == 1 else ''))
+
+        return tab
 
     def summarize_found(self, objects=None, start=None, stop=None):
         """Summarize found object database."""
@@ -457,6 +567,18 @@ class ZChecker(SBSearch):
             WHERE nightid=?
             ''', (exposures, quads, retrieved, nightid))
 
+        self.db.commit()
+
+        # make sure we do not add any duplicates
+        test = '''
+        SELECT count() FROM ztf WHERE pid=:pid
+        '''
+        new = [not bool(self.db.execute(test, {'pid': row[0]}).fetchone()[0])
+               for row in tab]
+        tab = tab[new]
+        exposures = len(np.unique(tab['expid']))
+        quads = len(tab)
+
         def obs_iterator(tab):
             for i in range(len(tab)):
                 row = tuple(tab[i].as_void())
@@ -470,11 +592,15 @@ class ZChecker(SBSearch):
                 row = tuple(tab[i].as_void())
                 jd_mid = float(row[1]) + row[2] / 86400 / 2
                 obsdate = Time(jd_mid, format='jd').iso[:-4]
-                ztf = (row[0], nightid, obsdate) + row[13:]
+                if isinstance(row[-1], bytes):
+                    maglim = None
+                else:
+                    maglim = float(row[-1])
+                ztf = (row[0], nightid, obsdate) + row[13:-1] + (maglim,)
                 yield ztf
 
         ztf_insert = '''
-        INSERT OR IGNORE INTO ztf VALUES (
+        INSERT INTO ztf VALUES (
           last_insert_rowid(),?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?
         )
         '''